+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/analysis_types.json
