+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/commitment_changes.sql
