+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/commitments.sql
