+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/consumption_beyond_commitments_standard.sql
