+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/consumption_through_commitments_enterprise_plus.sql
