+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/consumption_through_commitments_standard.sql
