+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/fulfillment_analysis.sql
