+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/reservation_timeline_analysis.sql
