+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/streaming_summary.sql
