+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/table_storage.sql
