+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/timeline_analysis.sql
