+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/workload_types.sql
