+++ src/databricks/labs/lakebridge/resources/assessments/bigquery/resources/write_api_summary.sql
